Add Death

+++ Assets/Scripts/Types/Interfaces.cs
@@ -5,22 +5,21 @@ namespace Landkreuzer.Types {
 	public interface IBeing {
 		BeingParameters BeingParameters { get; set; }
 		
-		
-		int Health { get; }
+		float Health { get; }
 		
 		/// <summary>
 		/// Decrease health
 		/// </summary>
 		/// <param name="value">Pure damage points before applying modificators</param>
 		/// <returns>Health after hurting</returns>
-		int Hurt(uint value);
+		float Hurt(float value);
 		
 		/// <summary>
 		/// Increase health 
 		/// </summary>
 		/// <param name="value">Pure heal points before applying modificators</param>
 		/// <returns>Health after healing</returns>
-		int Heal(uint value);
+		float Heal(float value);
 		
 		UnityEvent OnDead { get; }
 	}
